bdev: convert the type to avoid overflow
bdev_ch->io_outstanding - NOMEM_THRESHOLD_COUNT

can be negative, so change the type, then
we can make the comparision correct.

+++ lib/bdev/bdev.c
@@ -1751,8 +1751,8 @@ spdk_bdev_io_complete(struct spdk_bdev_io *bdev_io, enum spdk_bdev_io_status sta
 			 *  NOMEM_THRESHOLD_COUNT I/O to complete but for low queue
 			 *  depth channels we will instead wait for half to complete.
 			 */
-			bdev_ch->nomem_threshold = spdk_max(bdev_ch->io_outstanding / 2,
-							    bdev_ch->io_outstanding - NOMEM_THRESHOLD_COUNT);
+			bdev_ch->nomem_threshold = spdk_max((int64_t)bdev_ch->io_outstanding / 2,
+							    (int64_t)bdev_ch->io_outstanding - NOMEM_THRESHOLD_COUNT);
 			return;
 		}
 	}
